Schüler Kurse hinzufügen und löschen implementiert
Bug behoben beim Löschen von n:m-Tabellen (Transaktionsfehler beim Versuch, zwei vorher einem Schüler hinzugefügte Codes vom Schüler wieder zu entfernen. Grund: bei jedem Einfügen eines neuen Codes werden alle (!) Codes des Schülers in der Zwischentabelle upgedated. Bug? Abhilfe: Schüler vor löschen neu laden)

Bug behoben mit Löschen-Button (war enabled, wenn Tabellen leer)

Fehlerdialog für Transaktionsfehler anzeigen

+++ src/main/java/ch/metzenthin/svm/domain/commands/DeleteKursCommand.java
@@ -31,10 +31,10 @@ public class DeleteKursCommand extends GenericDaoCommand {
     public void execute() {
         KursDao KursDao = new KursDao(entityManager);
         Kurs KursToBeDeleted = Kurs.get(indexKursToBeDeleted);
-//        if (KursToBeDeleted.getSchueler().size() > 0) {
-//            result = Result.KURS_VON_SCHUELER_REFERENZIERT;
-//            return;
-//        }
+        if (KursToBeDeleted.getSchueler().size() > 0) {
+            result = Result.KURS_VON_SCHUELER_REFERENZIERT;
+            return;
+        }
         KursDao.remove(KursToBeDeleted);
         Kurs.remove(indexKursToBeDeleted);
         result = Result.LOESCHEN_ERFOLGREICH;
